출석체크 마무리 및 컴포넌트 쪼개기 중
